+++ MaxText/pyconfig.py
@@ -497,11 +497,9 @@ class _HyperParameters:
   def user_init(raw_keys):
     """Transformations between the config data and configs used at runtime"""
     if raw_keys["run_name"] == "":
-      raw_keys["run_name"] = os.environ.get("JOBSET_NAME")  # using XPK default
-      if raw_keys["run_name"] == "":
-        now = datetime.datetime.now()
-        timestamp = now.strftime("%Y-%m-%d-%H-%M")
-        raw_keys["run_name"] = f'{raw_keys["model_name"]}_{timestamp}'
+      now = datetime.datetime.now()
+      timestamp = now.strftime("%Y-%m-%d-%H-%M")
+      raw_keys["run_name"] = f'{raw_keys["model_name"]}_{timestamp}'
     run_name = raw_keys["run_name"]
     base_output_directory = raw_keys["base_output_directory"]
     if run_name:
