+++ MaxText/train.py
@@ -76,6 +76,9 @@ from ml_goodput_measurement import monitoring
 
 import MaxText as mt
 
+# Transformer = models.Transformer
+
+
 # pylint: disable=too-many-positional-arguments
 
 
@@ -574,73 +577,73 @@ def check_example_batch(config, example_batch):
     err.throw()
 
 
-# def setup_mesh_and_model(config, devices=None):
-#   """Set up the mesh and the model for training
+def setup_mesh_and_model(config, devices=None):
+  """Set up the mesh and the model for training
 
-#   Args:
-#     config
-#     devices
+  Args:
+    config
+    devices
 
-#   Returns:
-#     init_rng: RNG key
-#     writer: Summary writer for tensorboard
-#     checkpoint_manager: Orbax checkpointer
-#     state_mesh_annotations: the mesh annotations for the train state
-#     model:
-#     mesh:
-#     learning_rate_schedule:
-#     tx:
-#   """
+  Returns:
+    init_rng: RNG key
+    writer: Summary writer for tensorboard
+    checkpoint_manager: Orbax checkpointer
+    state_mesh_annotations: the mesh annotations for the train state
+    model:
+    mesh:
+    learning_rate_schedule:
+    tx:
+  """
 
-#   init_rng = random.PRNGKey(config.init_weights_seed)
-#   writer = max_utils.initialize_summary_writer(config.tensorboard_dir, config.run_name)
+  init_rng = random.PRNGKey(config.init_weights_seed)
+  writer = max_utils.initialize_summary_writer(config.tensorboard_dir, config.run_name)
 
-#   # Mesh definition
-#   devices_array = maxtext_utils.create_device_mesh(config, devices)
-#   mesh = Mesh(devices_array, config.mesh_axes)
+  # Mesh definition
+  devices_array = maxtext_utils.create_device_mesh(config, devices)
+  mesh = Mesh(devices_array, config.mesh_axes)
 
-#   # Model and Optimizer definition
-#   quant = quantizations.configure_quantization(config)
-#   model = Transformer(config, mesh, quant=quant)
-#   learning_rate_schedule = maxtext_utils.create_learning_rate_schedule(config)
-#   tx = optimizers.get_optimizer(config, learning_rate_schedule)
-#   logger = checkpointing.setup_checkpoint_logger(config)
-#   if config.enable_emergency_checkpoint:
-#     if config.use_replicator_service:
-#       checkpoint_manager = checkpointing.create_orbax_emergency_replicator_checkpoint_manager(
-#           config.local_checkpoint_directory,
-#           config.local_checkpoint_period,
-#           mesh,
-#       )
-#     else:
-#       abstract_state, _, _ = maxtext_utils.get_abstract_state(model, tx, config, init_rng, mesh, is_training=True)
-#       checkpoint_manager = checkpointing.create_orbax_emergency_checkpoint_manager(
-#           config.local_checkpoint_directory,
-#           config.checkpoint_dir,
-#           mesh,
-#           abstract_state,
-#           config.local_checkpoint_period,
-#           config.checkpoint_period,
-#           logger,
-#       )
-#   else:
-#     # TODO(b/368121306): Remove this once zarr3 support is plumbed on the backend
-#     use_ocdbt = config.checkpoint_storage_use_ocdbt
-#     use_zarr3 = config.checkpoint_storage_use_zarr3
-#     if config.enable_single_controller:
-#       use_ocdbt, use_zarr3 = False, False
-#     checkpoint_manager = checkpointing.create_orbax_checkpoint_manager(
-#         config.checkpoint_dir,
-#         config.enable_checkpointing,
-#         config.async_checkpointing,
-#         config.checkpoint_period,
-#         config.dataset_type,
-#         logger,
-#         use_ocdbt,
-#         use_zarr3,
-#     )
+  # Model and Optimizer definition
+  quant = quantizations.configure_quantization(config)
+  model = models.Transformer(config, mesh, quant=quant)
+  learning_rate_schedule = maxtext_utils.create_learning_rate_schedule(config)
+  tx = optimizers.get_optimizer(config, learning_rate_schedule)
+  logger = checkpointing.setup_checkpoint_logger(config)
+  if config.enable_emergency_checkpoint:
+    if config.use_replicator_service:
+      checkpoint_manager = checkpointing.create_orbax_emergency_replicator_checkpoint_manager(
+          config.local_checkpoint_directory,
+          config.local_checkpoint_period,
+          mesh,
+      )
+    else:
+      abstract_state, _, _ = maxtext_utils.get_abstract_state(model, tx, config, init_rng, mesh, is_training=True)
+      checkpoint_manager = checkpointing.create_orbax_emergency_checkpoint_manager(
+          config.local_checkpoint_directory,
+          config.checkpoint_dir,
+          mesh,
+          abstract_state,
+          config.local_checkpoint_period,
+          config.checkpoint_period,
+          logger,
+      )
+  else:
+    # TODO(b/368121306): Remove this once zarr3 support is plumbed on the backend
+    use_ocdbt = config.checkpoint_storage_use_ocdbt
+    use_zarr3 = config.checkpoint_storage_use_zarr3
+    if config.enable_single_controller:
+      use_ocdbt, use_zarr3 = False, False
+    checkpoint_manager = checkpointing.create_orbax_checkpoint_manager(
+        config.checkpoint_dir,
+        config.enable_checkpointing,
+        config.async_checkpointing,
+        config.checkpoint_period,
+        config.dataset_type,
+        logger,
+        use_ocdbt,
+        use_zarr3,
+    )
 
-#   return init_rng, writer, checkpoint_manager, mesh, model, learning_rate_schedule, tx
+  return init_rng, writer, checkpoint_manager, mesh, model, learning_rate_schedule, tx
 
 
 def setup_train_loop(config):
